add springql-config crate
