+++ springql-core/Cargo.toml
@@ -19,6 +19,7 @@ repository = "https://github.com/SpringQL/SpringQL"
 stub_web_console=[]
 
 [dependencies]
+springql-config = {version="0.18.0", features= ["default"], path="../springql-config"}
 anyhow = "1.0"
 thiserror = "1.0"
 serde = {version = "1.0", features = ["derive"], default-features = false}
@@ -38,10 +39,10 @@ reqwest = {version = "0.11", features = ["json", "blocking"], default-features =
 once_cell = "1.8"
 parking_lot = "0.12"
 time = {version="0.3.9", features = ["formatting", "parsing", "macros"]}
-
 socketcan = "1.7"
 
 [dev-dependencies]
+springql-config = {version="0.18.0", features= ["toml"], path="../springql-config"}
 springql-foreign-service = {path = "../foreign-service"}
 springql-test-logger = {path = "../test-logger"}
 pretty_assertions = "1.0"
