+++ springql/Cargo.toml
@@ -15,7 +15,8 @@ readme = "../README.md"
 repository = "https://github.com/SpringQL/SpringQL"
 
 [dependencies]
-springql-core =  { version = "0.18.1", path="../springql-core"}
+springql-config = {version="0.18.0", features= ["toml"], path="../springql-config"}
+springql-core =  { version = "0.18.0", path="../springql-core"}
 
 [dev-dependencies]
 springql-foreign-service = {path = "../foreign-service"}
